requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
django-filter==23.3
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
django-cors-headers==4.3.1
whitenoise==6.6.0
daphne==4.0.0
psycopg2-binary==2.9.10
django-environ==0.11.2
dnspython==2.3.0
django-redis==5.4.0
gunicorn==21.2.0
supervisor==4.2.5
pillow==11.3.0
celery==5.3.4
djangorestframework-simplejwt==5.3.0